Api call and search feature

+++ app/src/main/java/com/balakumar/weatherapplication/models/Current.kt
@@ -3,7 +3,7 @@ package com.balakumar.weatherapplication.models
 data class Current(
     val apparent_temperature: Double,
     val interval: Int,
-    val rain: Int,
+    val rain: Double,
     val relative_humidity_2m: Int,
     val temperature_2m: Double,
     val time: String,
